Creado el menu de Pausa en el que eh añadido dos botones uno para la musica del juego y otro para los efectos de sonido del juego

+++ src/main/java/proyectoFinal/SuperSteveBros/entities/Entity.java
@@ -28,7 +28,7 @@ public abstract class Entity {
 		root.getChildren().add(fxRect);
 	}
 
-	protected void iniHitbox(float x, float y, float width, float height) {
+	protected void iniHitbox(float x, float y, int width, int height) {
 		hitBox = new Rectangle2D.Float(x, y, width, height); 
 		fxRect = new Rectangle((int) hitBox.x, (int) hitBox.y, (int) hitBox.width, (int) hitBox.height);
 		fxRect.setFill(Color.TRANSPARENT);
